Refactor HomePage component to utilize custom hook for state management and logic
